# Prune backtraces similar to RUST_BACKTRACE=1 logic