@RalfJung 2022-02-20
> It might be nice to have the "some details were omitted" note and a fallback to a setting where we print everything just in case this logic goes sideways, but

>    I'm not sure where to put the note

You can just emit a final `note` after all the call frames saying that some frames were omitted.

>   MIRI_BACKTRACE, RUST_BACKTRACE, and RUSTC_CTFE_BACKTRACE already do something else. Should we repurpose or add on to the semantics of MIRI_BACKTRACE?

Those env vars are all for backtraces of the host interpreter/compiler. I feel like to affect the backtraces for a diagnostic, a `-Zmiri` flag would make more sense.