@RalfJung 2022-02-20
> A bit of before-and-after from cargo miri run:

Ah, very nice. :)
Kind of a shame that we still get that `call_once - shim`... I wonder if we should skip those shims when printing short backtraces?