@saethlin 2022-02-20
> Kind of a shame that we still get that `call_once - shim`... I wonder if we should skip those shims when printing short backtraces?

Originally I was keeping them in for symmetry with `RUST_BACKTRACE` but if you also think they're noise, that's more than enough for me.