@RalfJung 2022-02-20
> Originally I was keeping them in for symmetry with RUST_BACKTRACE but if you also think they're noise, that's more than enough for me.

I was thinking we could actually skip them in general, not just the particular one that wraps `main`. But we can leave that for a future PR.