@RalfJung 2022-02-20
This looks great, thanks a lot! In fact this might allow us to switch to ui tests for the compile-fail test suite, since the output of the backtraces from failing tests should be a lot more stable after this change.